+++ interpreter/variable/error.go
@@ -211,11 +211,6 @@ func (v *ErrorScopeVariables) Set(s context.Scope, name, operator string, val va
 		}
 		return nil
 	case OBJ_RESPONSE:
-		// If synthetic response has already been set, ignore this assignment.
-		if v.ctx.HasSyntheticResponse {
-			return nil
-		}
-
 		if err := doAssign(v.ctx.ObjectResponse, operator, val); err != nil {
 			return errors.WithStack(err)
 		}